+++ CloudShell.Providers.Applications/ApplicationResourceService.cs
@@ -45,6 +45,7 @@ public sealed partial class ApplicationResourceService(
     IHostScopedResourceCleanupProvider,
     IApplicationResourceManagementOperations,
     IApplicationResourceConfigurationOperations,
+    IApplicationResourceRegistrationOperations,
     IApplicationResourceRunningStateOperations,
     IApplicationResourceDefinitionSource,
     IApplicationResourceProcedureOperations,